+++ R/06_LRA.R
@@ -77,7 +77,7 @@ LRA.default <- function(U, na = NULL, Z = NULL, w = NULL, ...) {
 #' @param maxiter Maximum number of iterations for estimation. Default is 100.
 #' @param BIC.check For binary data with SOM method only. If TRUE, convergence is checked using BIC values. Default is FALSE.
 #' @param seed For binary data with SOM method only. Random seed for reproducibility.
-#' @param verbose Logical; if TRUE, displays detailed progress during estimation. Default is TRUE.
+#' @param verbose Logical; if TRUE, displays detailed progress during estimation. Default is FALSE.
 #' @param beta1 Beta distribution parameter 1 for prior density of rank reference matrix (GTM method only). Default is 1.
 #' @param beta2 Beta distribution parameter 2 for prior density of rank reference matrix (GTM method only). Default is 1.
 #' @param conf Confirmatory IRP matrix (items x nrank) for test equating.