+++ R/15_Biclustering_nominal.R
@@ -12,7 +12,7 @@
 #'   * A field membership profile matrix (items × fields) with 0/1 values
 #'   * NULL (default) for exploratory analysis where field memberships are estimated
 #' @param maxiter Maximum number of EM algorithm iterations. Default is 100.
-#' @param verbose Logical; if TRUE, displays progress during estimation. Default is TRUE.
+#' @param verbose Logical; if TRUE, displays progress during estimation. Default is FALSE.
 #' @param alpha Dirichlet distribution concentration parameter for prior density of field reference probabilities. Default is 1.
 #' @param ... Additional arguments passed to specific methods.
 #' #'
@@ -29,7 +29,7 @@ Biclustering.nominal <- function(U,
                                  conf_class = NULL,
                                  mic = FALSE,
                                  maxiter = 100,
-                                 verbose = TRUE,
+                                 verbose = FALSE,
                                  alpha = 1, ...) {
   tmp <- U
   tmp$Q <- remap_category_codes(tmp$Q)