+++ R/17_Biclustering_nominal_IRM.R
@@ -31,9 +31,10 @@
 #' @export
 Biclustering_IRM.nominal <- function(U,
                                      gamma_c = 1, gamma_f = 1, alpha = 1,
-                                     max_iter = 100, stable_limit = 5,
+                                     maxiter = 100, stable_limit = 5,
                                      minSize = 20, EM_limit = 20,
-                                     seed = 123, verbose = TRUE, ...) {
+                                     seed = 123, verbose = FALSE, ...) {
+  maxiter <- resolve_deprecated_max_iter(maxiter, list(...))
   tmp <- U
   tmp$Q <- remap_category_codes(tmp$Q)
 
@@ -91,7 +92,7 @@ Biclustering_IRM.nominal <- function(U,
   gibbs <- irm_gibbs_core(
     Uq = Uq, Z = tmp$Z, cls01 = cls01, fld01 = fld01,
     gamma_c = gamma_c, gamma_f = gamma_f, alpha_vec = alpha_vec,
-    max_iter = max_iter, stable_limit = stable_limit, verbose = verbose
+    max_iter = maxiter, stable_limit = stable_limit, verbose = verbose
   )
 
   cls01 <- gibbs$cls01